While importing a Daisy 3 /Daisy 2/ Epub 3 project 96000 Hz and 48000 Hz sample rate will work. Support for 96000 Hz and 48000 Hz added.

+++ Obi/Obi/ImportExport/DAISY3_ObiImport.cs
@@ -73,7 +73,7 @@ namespace Obi.ImportExport
             m_Presentation = m_Session.Presentation;
 
             PCMFormatInfo pcmFormat = m_Presentation.MediaDataManager.DefaultPCMFormat.Copy();
-            pcmFormat.Data.SampleRate = (ushort)m_audioProjectSampleRate;
+            pcmFormat.Data.SampleRate = (uint)m_audioProjectSampleRate;
             m_Presentation.MediaDataManager.DefaultPCMFormat = pcmFormat;
 
             m_Presentation.MediaDataManager.EnforceSinglePCMFormat = true;
